Add native Adobe module
